move upload test file to baseline
